Merge of APS/HUD with Latest Build
Nearly all of the features from my builds have been integrated with the
main build:
Rudimentary HUD system that displays entity statistics.
Three buffs and one debuff that all work.
New scripts: Architect and Hunter's Momemtum. Architect is the child of
Character and is attached to the Character GameObject (we should rename
it to the actual Character). Hunter's Momentum is an attribute trait for
the Architect that runs when the Architect has at least one point in
Marksman.
Significant changes were made to the scripts shown on the right. Expand
them to read the changes in code.
Updated using Unity 5.0.1.