Effect Classes V00002
Added crippled and swiftness classes, marking the complete replacement
of effect on the unit class with their own individual classes.

I'll be making a few passes on code that interacts with effects, and add
one or more new effects to cover all anticipated mechanics, like frozen
(stronger than cripple and opposite of haste), and enfeebled (opposite
of empowered).